bot/scripts/requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
requests>=2.31.0
psycopg2-binary>=2.9.9
deep-translator>=1.11.4
tqdm>=4.66.1
jsonschema>=4.20.0
aiohttp>=3.9.0
aiohappyeyeballs>=2.5.0
aiosignal>=1.4.0
attrs>=23.2.0
frozenlist>=1.4.0
multidict>=6.0.0
propcache>=0.2.0
yarl>=1.9.0
psutil>=7.0.0
argparse>=1.4.0
backoff>=2.2.1
psycopg2-pool>=1.1
flask>=2.3.3
apscheduler>=3.10.1